+++ libsolidity/formal/BMC.h
@@ -60,6 +60,7 @@ public:
 	BMC(
 		smt::EncodingContext& _context,
 		langutil::UniqueErrorReporter& _errorReporter,
+		langutil::UniqueErrorReporter& _unsupportedErrorReporter,
 		std::map<h256, std::string> const& _smtlib2Responses,
 		ReadCallback::Callback const& _smtCallback,
 		ModelCheckerSettings _settings,