+++ libsolidity/formal/ModelChecker.h
@@ -89,6 +89,12 @@ private:
 	/// to m_errorReporter at the end of the analysis.
 	langutil::UniqueErrorReporter m_uniqueErrorReporter;
 
+	/// Used by SMTEncoder, BMC and CHC to accumulate unsupported
+	/// warnings and avoid duplicates.
+	/// This is local to ModelChecker, so needs to be appended
+	/// to m_errorReporter at the end of the analysis.
+	langutil::UniqueErrorReporter m_unsupportedErrorReporter;
+
 	ModelCheckerSettings m_settings;
 
 	/// Stores the context of the encoding.